Move and rename file to LocoFixedVector.hpp
