+++ src/OpenLoco/Core/LocoFixedVector.hpp
