+++ src/OpenLoco/StationManager.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "LocoFixedVector.h"
+#include "Core/LocoFixedVector.hpp"
 #include "Station.h"
 #include <array>
 #include <cstddef>
