+++ src/OpenLoco/TownManager.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "LocoFixedVector.h"
+#include "Core/LocoFixedVector.hpp"
 #include "Town.h"
 #include <array>
